Move compare_NIPA_to_IOT into the nowcasting package
It is the NIPA side of the nowcast reconciliation work and nothing outside
nowcasting uses it: the only importers were value_added_control_totals.py and
plan.md. Sitting a level up in analysis/ implied a general-purpose utility that
it is not.

All 13 files moved as git renames, so history follows. Rewrites:

- bedrock.analysis.compare_NIPA_to_IOT -> bedrock.analysis.nowcasting.compare_NIPA_to_IOT
  across the package's own modules, its tests, its two examples, and
  value_added_control_totals.py
- the pyproject per-file ruff exemption for loaders.py (PLC0415)
- plan.md's link, which now points at a sibling subfolder rather than up and
  across
- the moved README's ../.. links, which needed one more hop now that the
  package sits one level deeper

Verified: 148 analysis tests pass (83 of them this package's own), ruff, black
and mypy clean across all 17 source files, both CLIs run, and no broken
relative links anywhere under nowcasting/.
