+++ bedrock/analysis/nowcasting/compare_NIPA_to_IOT/__init__.py
@@ -5,7 +5,7 @@ matching and reports the weak links rather than demanding an exact crosswalk, so
 a new dataset can be sanity-checked against BEA in a few lines.  When a number
 needs to be defensible, promote the comparison to a real crosswalk instead.
 
-    from bedrock.analysis.compare_NIPA_to_IOT import (
+    from bedrock.analysis.nowcasting.compare_NIPA_to_IOT import (
         bea_matrix_row, compare, nipa_flat_table
     )
 
@@ -18,15 +18,19 @@ needs to be defensible, promote the comparison to a real crosswalk instead.
     )
 """
 
-from bedrock.analysis.compare_NIPA_to_IOT.compare import ROLLUPS, Comparison, compare
-from bedrock.analysis.compare_NIPA_to_IOT.hierarchy import (
+from bedrock.analysis.nowcasting.compare_NIPA_to_IOT.compare import (
+    ROLLUPS,
+    Comparison,
+    compare,
+)
+from bedrock.analysis.nowcasting.compare_NIPA_to_IOT.hierarchy import (
     RESIDUAL_MARKERS,
     RESIDUAL_TOKENS,
     markers_for,
     split_residual,
     token_relation,
 )
-from bedrock.analysis.compare_NIPA_to_IOT.loaders import (
+from bedrock.analysis.nowcasting.compare_NIPA_to_IOT.loaders import (
     BEA_MATRICES,
     FRAMEWORK_NAMES,
     UNAVAILABLE_MATRICES,
@@ -45,8 +49,8 @@ from bedrock.analysis.compare_NIPA_to_IOT.loaders import (
     table_series,
     where_is,
 )
-from bedrock.analysis.compare_NIPA_to_IOT.matching import Alignment, align
-from bedrock.analysis.compare_NIPA_to_IOT.series import (
+from bedrock.analysis.nowcasting.compare_NIPA_to_IOT.matching import Alignment, align
+from bedrock.analysis.nowcasting.compare_NIPA_to_IOT.series import (
     LabeledSeries,
     normalize_code,
     normalize_name,
